+++ src/export/export.js
@@ -2,46 +2,10 @@
 
 const { promises: jsonld } = require('jsonld')
 
-const { camelize, omit } = require('./util')
+const { shorten } = require('./utils')
 const { ITEM } = require('../constants/type')
 const { TEMPLATE } = require('../constants/ontology')
-const { getLabel } = require('./ontology')
 const { entries, values, keys } = Object
-
-
-
-function propertyLabel(property, props, template) {
-  var label, field
-  try {
-    if (template) {
-      field = template.fields.find(f => f.property === property)
-      label = field && field.label
-    }
-    if (!label) {
-      label = props[property] && props[property].label
-    }
-    if (!label) {
-      label = getLabel(field.id)
-    }
-    return label
-  } catch (_) { return label }
-}
-
-function shortenLabel(label) {
-  return camelize(
-    label
-    .toLowerCase()
-    .trim()
-    .normalize('NFD')                // normalize unicode
-    .replace(/[\u0300-\u036f]/g, '') // remove accents, ligatures
-    .replace(/[^a-zA-Z0-9]+/g, ' ')  // remove non-alphanumeric
-  ).replace('_', ' ')                // remove _
-}
-
-function shorten() {
-  const label = propertyLabel(...arguments)
-  if (label) return shortenLabel(label)
-}
 
 function makeContext(metadata, template, props) {
   const result = {
@@ -124,45 +88,6 @@ async function groupedByTemplate(resources, props = {}) {
   return results
 }
 
-//////// Import
-
-class ParseError extends Error {
-  constructor(obj, ...args) {
-    super(...args)
-    Error.captureStackTrace(this, ParseError)
-
-    this.details = JSON.stringify(obj, null, 2)
-  }
-}
-
-async function itemFromLD(obj) {
-  let metadata, type, templateID
-  try {
-    const [expanded] = await jsonld.expand(obj)
-    type = expanded['@type'][0]
-    templateID = expanded[TEMPLATE.TYPE][0]['@id']
-    metadata = omit(expanded, ['@type', TEMPLATE.TYPE])
-  } catch (e) {
-    throw new ParseError(obj, 'Could not parse jsonld object')
-  }
-
-  // convert metadata to a format supported by `mod.item.create`
-  let md = {}
-  for (let property in metadata) {
-    const [prop] = metadata[property]
-    md[property] =  {
-      type: prop['@type'],
-      text: prop['@value']
-    }
-  }
-
-  return { type, templateID, metadata: md }
-}
-
 module.exports = {
-  shortenLabel,
-  propertyLabel,
-  itemFromLD,
-  ParseError,
   groupedByTemplate
 }
