+++ src/export/index.js
@@ -1,0 +1,10 @@
+'use strict'
+
+const { groupedByTemplate } = require('./export')
+const { itemFromLD, ParseError } = require('./import')
+
+module.exports = {
+  groupedByTemplate,
+  itemFromLD,
+  ParseError
+}
