+++ src/export/utils.js
@@ -1,0 +1,43 @@
+'use strict'
+
+const { camelize } = require('../common/util')
+const { getLabel } = require('../common/ontology')
+
+function propertyLabel(property, props, template) {
+  var label, field
+  try {
+    if (template) {
+      field = template.fields.find(f => f.property === property)
+      label = field && field.label
+    }
+    if (!label) {
+      label = props[property] && props[property].label
+    }
+    if (!label) {
+      label = getLabel(field.id)
+    }
+    return label
+  } catch (_) { return label }
+}
+
+function shortenLabel(label) {
+  return camelize(
+    label
+    .toLowerCase()
+    .trim()
+    .normalize('NFD')                // normalize unicode
+    .replace(/[\u0300-\u036f]/g, '') // remove accents, ligatures
+    .replace(/[^a-zA-Z0-9]+/g, ' ')  // remove non-alphanumeric
+  ).replace('_', ' ')                // remove _
+}
+
+function shorten() {
+  const label = propertyLabel(...arguments)
+  if (label) return shortenLabel(label)
+}
+
+module.exports = {
+  shortenLabel,
+  propertyLabel,
+  shorten
+}
